gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2